Add files via upload
There are three files in this directory.
The main file is FreqTURBOGraph_ENGULx.vi. 
It calls up the unit conversion file (mili, micro, etc.) Convert.vi as well as the temperature meter file ULxTemperature.vi.
The last file is used to measure temperature using Measurement Computing USB-TEMP-AI temperature meter. 
If a different thermometer is used, it must be replaced.